tests/requirements.txt
selenium==4.23.1
requests==2.32.3
Pillow==10.4.0
pytest==8.3.2
pytest-html==4.1.1
pytest-xdist==3.6.1
